backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
camelot-py==0.11.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
dataclasses-json==0.6.6
DAWG-Python==0.7.2
deepdiff==7.0.1
distro==1.9.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
emoji==2.12.1
et-xmlfile==1.1.0
filetype==1.2.0
frozenlist==1.4.1
gunicorn==21.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.75
lxml==5.2.2
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.32.0
opencv-python==4.10.0.82
openpyxl==3.1.3
ordered-set==4.1.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
playwright==1.44.0
plotly==5.22.0
prompt_toolkit==3.0.46
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pyee==11.1.0
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
pypdf==4.2.0
PyPDF2==2.9.0
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.18.1
ruwordnet==0.0.6
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.52
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
unstructured==0.14.5
unstructured-client==0.23.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
html5lib
# chatgpt-proxy
kaleido==0.2.1
xhtml2pdf==0.2.16
python-docx==1.1.2
xlsxwriter==3.2.0
openpyxl==3.1.3
# gh-md-to-html==1.21.2
markdown==3.6
